+++ packages/blocks/core/js/wordpress/avatar/test/functionality.blocks-1.e2e.cy.js
@@ -181,7 +181,25 @@ describe('Avatar Block', () => {
 			});
 
 		//
-		// Assert front end
+		// 2.0. Settings tab
+		//
+		cy.getByDataTest('settings-tab').click();
+
+		cy.get('.block-editor-block-inspector').within(() => {
+			// the range slide for image size should be hidden
+			cy.get(
+				'.components-panel__body .components-range-control .components-base-control__label'
+			).should('not.visible');
+			cy.get('.components-panel__body .components-range-control').should(
+				'not.visible'
+			);
+			cy.get('.components-panel__body .components-toggle-control').should(
+				'be.visible'
+			);
+		});
+
+		//
+		// 3.0 Assert front end
 		//
 		savePage();
 		redirectToFrontPage();